// src/app/modules/billing/discount/models/discount.model.ts
export interface DiscountData {
   billId: number;
   billNum: number;
   batchNum: string;
   custName: string;
   blngLevel: string;
   blngLevelCode: string;
   sfxCode:string;
   msaCode: string;
   amount: string;
   discSign: string;
   discType: string;
   value: string;
   discountVal: string;
   revisedBillAmount: string;
   revisedGBillAmount: string;
}